Force auto rotation in DayDreams
There's a bug in Android DayDreams which causes the screen
orientation to change to portrait as soon as the lock
screen is activated. Optionally users can now choose to force
the auto rotation of the screen. The app will start its activity when
the DayDream starts, which is not affected by the bug.